Owner v2.0
I moved `ArrayList<Property> propertiesOwned` to Owner as I think it makes more sense there as each Owner has an arraylist of the properties they own. 

Because of this I have added an `addProperties` method to it and a `getProperties` method. 

Finally I added a `toString` method to give the properties of the owner